requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.3.0
chromadb>=0.4.0
langchain-chroma>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.30.0
langsmith>=0.1.0
sqlite-utils>=3.35.0 